Split Gui into ListsGui and MainDisplayGui.
